auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography
python-jose
pytest
httpx


